# lib/httpServer/index.js
const express = require('express')
const bodyParser = require('body-parser')
const swaggerUi = require('swagger-ui-express')
const R = require('ramda')
const swaggerDocument = require('./swagger.json')
const Block = require('../blockchain/block')
const Transaction = require('../blockchain/transaction')
const TransactionAssertionError = require('../blockchain/transactionAssertionError')
const BlockAssertionError = require('../blockchain/blockAssertionError')
const HTTPError = require('./httpError')
const ArgumentError = require('../util/argumentError')
const CryptoUtil = require('../util/cryptoUtil')

class HttpServer {
  constructor (node, blockchain, operator, miner) {
    this.app = express()

    const projectWallet = (wallet) => {
      return {
        id: wallet.id,
        addresses: R.map((keyPair) => {
          return keyPair.publicKey
        }, wallet.keyPairs)
      }
    }

    this.app.use(bodyParser.json())

    this.app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument))

    this.app.get('/blockchain/blocks', (req, res) => {
      res.status(200).send(blockchain.getAllBlocks())
    })

    this.app.get('/blockchain/blocks/latest', (req, res) => {
      let lastBlock = blockchain.getLastBlock()
      if ( ! lastBlock) throw new HTTPError(404, 'Last block not found')

      res.status(200).send(lastBlock)
    })

    this.app.put('/blockchain/blocks/latest', (req, res) => {
      let requestBlock = Block.fromJson(req.body)
      let result = node.checkReceivedBlock(requestBlock)

      if ( ! result) res.status(200).send('Requesting the blockchain to check.')
      else if (result) res.status(200).send(requestBlock)
      else throw new HTTPError(409, 'Blockchain is update.')
    })

    this.app.get('/blockchain/blocks/:hash([a-zA-Z0-9]{64})', (req, res) => {
      let blockFound = blockchain.getBlockByHash(req.params.hash)
      if ( ! blockFound) throw new HTTPError(404, `Block not found with hash '${req.params.hash}'`)

      res.status(200).send(blockFound)
    })

    this.app.get('/blockchain/blocks/:index', (req, res) => {
      let blockFound = blockchain.getBlockByIndex(req.params.index)
      if ( ! blockFound) throw new HTTPError(404, `Block not found with index '${req.params.index}'`)

      res.status(200).send(blockFound)
    })

    this.app.get('/blockchain/blocks/transactions/:transactionId([a-zA-Z0-9]{64})', (req, res) => {
      let transactionFromBlock = blockchain.getTransactionFromBlocksById(req.params.transactionId)
      if ( ! transactionFromBlock) throw new HTTPError(404, `Transaction '${req.params.transactionId}' not found in any block`)

      res.status(200).send(transactionFromBlock)
    })

    this.app.get('/blockchain/transactions', (req, res) => {
      res.status(200).send(blockchain.getAllPendingTransactions())
    })

    this.app.post('/blockchain/transactions', (req, res) => {
      let requestTransaction = Transaction.fromJson(req.body)
      let transactionFound = blockchain.getPendingTransactionById(requestTransaction.id)

      if ( ! transactionFound) throw new HTTPError(409, `Transaction '${requestTransaction.id}' already exists`)

      try {
        let newTransaction = blockchain.addTransaction(requestTransaction)
        res.status(201).send(newTransaction)
      } catch (ex) {
        if (ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, requestTransaction, ex)
        else throw ex
      }
    })

    this.app.get('/blockchain/transactions/unspent', (req, res) => {
      res.status(200).send(blockchain.getUnspentTransactionsForAddress(req.query.address))
    })

    this.app.get('/operator/wallets', (req, res) => {
      let wallets = operator.getWallets()

      let projectedWallets = R.map(projectWallet, wallets)

      res.status(200).send(projectedWallets)
    })

    this.app.post('/operator/wallets', (req, res) => {
      let password = req.body.password
      if (R.match(/\w+/g, password).length <= 4) throw new HTTPError(400, 'Password must contain more than 4 words')

      let newWallet = operator.createWalletFromPassword(password)

      let projectedWallet = projectWallet(newWallet)

      res.status(201).send(projectedWallet)
    })

    this.app.get('/operator/wallets/:walletId', (req, res) => {
      let walletFound = operator.getWalletById(req.params.walletId)
      if ( ! walletFound) throw new HTTPError(404, `Wallet not found with id '${req.params.walletId}'`)

      let projectedWallet = projectWallet(walletFound)

      res.status(200).send(projectedWallet)
    })

    this.app.post('/operator/wallets/:walletId/transactions', (req, res) => {
      let walletId = req.params.walletId
      let password = req.headers.password

      if ( ! password) throw new HTTPError(401, 'Wallet\'s password is missing.')
      let passwordHash = CryptoUtil.hash(password)

      try {
        if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`)

        let newTransaction = operator.createTransaction(walletId, req.body.fromAddress, req.body.toAddress, req.body.amount, req.body.changeAddress)

        newTransaction.check()

        let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction))
        res.status(201).send(transactionCreated)
      } catch (ex) {
        if (ex instanceof ArgumentError || ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, walletId, ex)
        else throw ex
      }
    })

    this.app.get('/operator/wallets/:walletId/addresses', (req, res) => {
      let walletId = req.params.walletId
      try {
        let addresses = operator.getAddressesForWallet(walletId)
        res.status(200).send(addresses)
      } catch (ex) {
        if (ex instanceof ArgumentError) throw new HTTPError(400, ex.message, walletId, ex)
        else throw ex
      }
    })

    this.app.post('/operator/wallets/:walletId/addresses', (req, res) => {
      let walletId = req.params.walletId
      let password = req.headers.password

      if ( ! password) throw new HTTPError(401, 'Wallet\'s password is missing.')
      let passwordHash = CryptoUtil.hash(password)

      try {
        if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`)

        let newAddress = operator.generateAddressForWallet(walletId)
        res.status(201).send({ address: newAddress })
      } catch (ex) {
        if (ex instanceof ArgumentError) throw new HTTPError(400, ex.message, walletId, ex)
        else throw ex
      }
    })

    this.app.get('/operator/wallets/:walletId/addresses/:addressId/balance', (req, res) => {
      let walletId = req.params.walletId
      let addressId = req.params.addressId

      try {
        let balance = operator.getBalanceForWalletAddress(walletId, addressId)
        res.status(200).send({ balance: balance })
      } catch (ex) {
        if (ex instanceof ArgumentError) throw new HTTPError(400, ex.message, { walletId, addressId }, ex)
        else throw ex
      }
    })

    this.app.get('/node/peers', (req, res) => {
      res.status(200).send(node.peers)
    })

    this.app.post('/node/peers', (req, res) => {
      let newPeer = node.connectToPeer(req.body)
      res.status(201).send(newPeer)
    })

    this.app.get('/node/transactions/:transactionId([a-zA-Z0-9]{64})/confirmations', (req, res) => {
      node.getConfirmations(req.params.transactionId)
                .then((confirmations) => {
                  res.status(200).send({ confirmations: confirmations })
                })
    })

    this.app.post('/miner/mine', (req, res, next) => {
      miner.mine(req.body.rewardAddress)
                .then((newBlock) => {
                  newBlock = Block.fromJson(newBlock)
                  blockchain.addBlock(newBlock)
                  res.status(201).send(newBlock)
                })
                .catch((ex) => {
                  if (ex instanceof BlockAssertionError && ex.message.includes('Invalid index')) next(new HTTPError(409, 'A new block were added before we were able to mine one'), null, ex)
                  else next(ex)
                })
    })

    this.app.use(function (err, req, res, next) {  // eslint-disable-line no-unused-vars
      if (err instanceof HTTPError) res.status(err.status)
      else res.status(500)
      res.send(err.message + (err.cause ? ' - ' + err.cause.message : ''))
    })
  }

  listen (host, port) {
    return new Promise((resolve, reject) => {
      this.server = this.app.listen(port, host, (err) => {
        if (err) reject(err)
        console.info(`Listening http on port: ${this.server.address().port}, to access the API documentation go to http://${host}:${this.server.address().port}/api-docs/`)
        resolve(this)
      })
    })
  }

  close () {
    if (this.server) {
      this.server.close()
    }
  }
}

module.exports = HttpServer